requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudevents==1.2.0
deprecation==2.1.0
feedparser==6.0.2
Flask==1.1.4
functions-framework==2.1.3
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sgmllib3k==1.0.0
six==1.16.0
tweepy==3.10.0
uritemplate==3.0.1
urllib3==1.26.4
watchdog==1.0.2
Werkzeug==1.0.1
